Added webui buttons and endpoints for manual On/Off. Added check for server health response. General refactoring. Cleaned up "check args" in main. All args checking is now in the same place. Cleaned up "config path" in main. Function moved to config.rs Renamed webserver.rs to webui.rs

+++ src/price.rs
@@ -7,7 +7,6 @@ use serde_json::Value;
 use std::fs::File;
 use std::io::Write;
 use std::time::Duration as TimeDuration;
-use std::usize;
 use std::env;
 
 use crate::structs;
@@ -141,7 +140,7 @@ pub fn ratio_price(json: &Value, currency: &str, ratio: f64) -> Option<f64> {
         .collect();
 
     prices.sort_by(|a, b| a.partial_cmp(b).unwrap());
-    prices.get(index as usize).copied()
+    prices.get(index).copied()
 }
 
 fn extract_valid_price(obj: &Value, currency: &str) -> Option<f64> {
